--dry-run and diffs
Support dry-run mode where we process the sources and build entries, but
don't make any modifications to actual catalog data. When in dry-run,
we'll output diffs of everything we might change, so people can preview
before they accept the changes.

In future we may want to separate the concept of dry-running from diffs,
but for now we combine the two (just a bit simpler).