Alias support and Backstage is namespace aware
Our support for aliases was a bit borked, so this commit fixes it by
accepting either a string or an []string from any of the alias CEL
expressions and combining them together.

Then we change Backstage so:
- The external ID is namespace/name, supporting people who have many
  resources with the same name
- Use the name as an alias to ensure we continue to match for those
  without duplicates

The effect is that Backstage users can get all the benefits of
namespace'ing but only if they want it, while others who use a single
namespace shouldn't notice the difference.