Show catalog type name on creation error (#43)
The catalog-importer should show the catalog type name if it attempts to create a catalog type that already exists, to be more useful. The catalog-importer does check for existing catalog types before attempting to create new ones from an importer config, but that list of existing catalog types ignores any catalog types that are not managed by an importer or managed by a different importer, so this might explain why that could have happened.