catalog-importer import
For those who want to import an existing file of data or do a one-time
import, this command can help achieve that.

I've tested it against a CSV export of a Notion database. I asked the
import to actually run a sync but it's equally valid not to, and to rely
on the import command to generate the importer.jsonnet config that you
add to your broader sync config and use whatever source files you run
the import against to be the source of the data from thereon out.

The command I was using is this:

```
$ catalog-importer import \
  --local-file tmp/documents.csv \
  --name Documents \
  --description 'Documents from a Notion database' \
  --type-name 'Custom["Documents"]' \
  --source-external-id 'Name' \
  --source-name 'Name' \
  --run-sync \
  --run-sync-dry-run
```

This commit includes two broader changes which are:
- Setting the top-level scope of a CEL expression to the underscore
  character
- Providing support for CSV input files