exec source: Dump stdout upon failure
Some commands which you may want to use in an importer configuration are
not well behaved, and send output which constitutes the error condition
to `stdout`, rather than `stderr`. An example of this is `bq`.

This effectively squashes the internal error message and makes it much
harder to debug the issue than it should be.

To remedy this, in the cases where the command fails we dump at most
1KiB of the output back into stderr.