exec source: Pipe stdout to terminal
Some commands which you may want to use in an importer configuration are
not well behaved, and send output which constitutes the error condition
to `stdout`, rather than `stderr`. An example of this is `bq`.

This effectively squashes the internal error message and makes it much
harder to debug the issue than it should be.

To remedy this, stream _both_ `stdout` and `stderr` to the parent
terminal.