Reject unknown keys from importer config
This tripped up a customer where they typo'd `outputs` as `output` and
wondered why their config did nothing.

I don't love the output when we fail to load the config file right now
(it can be improved a lot!) but this commit makes sure we fail with an
error like so:

```
main: error: loading config: parsing config: json: unknown field "invalid_key"
```

In the situation where we don't recognise keys in the config.