Increase permissiveness of JS expression evaluation
Previously, our CEL expression evaluation would log errors out and continue if it failed. We're moving back to that level of permissiveness with our JS expression evaluation. The expected behaviour is that if JS eval fails for a given source, the error will be logged (in DEBUG mode), and the importer will continue without hard-failing.