Update goreleaser to publish tag of current major version (#91)
Part of work to make it easier to manage github types externally. 

We want to be able to configure github workflows to pull the latest tag
of a specific major version (i.e 2.X.X), rather than having to specify
either latest (which might break if major version shift), or specific
version (which would require users to update their configs if we made
minor changes).