Don't set attributes that aren't in the input
When an attribute wasn't in the input data, we weren't handling that
correctly and would set the value of the attribute in the catalog to the
Go empty value (for example, we'd set a number attribute to 0 if that
attribute wasn't set in the input for an entry).

This has been fixed by distinguishing when the value we get back from
Otto is null or undefined, and not including that attribute.