Support backlink attributes via the importer
We're introducing backlink attributes, which allow you to say 'find all the features that this team owns' or similar, based on another catalog type.

This is challenging for the importer as this means attributes in an importer config can reference each other. e.g. if I'm creating both Teams and Features in my catalog, I need to wait until Features exists before creating the backlink attribute in Teams.

To resolve this, we split the UpdateTypeSchema process into two parts:
1. Create/Update all the types and their schemas
2. Create any new backlinks that don't exist yet (as the other attribute must now be present).

Note that to make the code sane, I've pulled the 'if dryRun' up one level as dryRun doesn't have any of these concerns.