Default to a 1 second backoff when hitting 429s
When importing a large amount of new entries, or deleting them we're
hitting against the general API limit.

Using a pool improves our performance but we still seem to see the odd
failure where the 10 routines are fighting one another enough that we
fail after X retries (tried up to 10 and still saw issues).

The big fix that worked here was to ensure we backoff for a minimum of 1
second at least (anything smaller, or even negative means we still do 1
now).

This results in the rate we submit results peaking at several hundred
requests per second until we hit our per min API rate limit, then we are
at the mercy of the token bucket refill - essentially doing max we can
as our rate limit refills.

This slows down a little, but it's in theory the best we can do per the
rate limit and we're a better citizen talking to the API.