Support configuring the page size for listing catalog entries (#151)
In cases where listing entries can be expensive, it's useful to be able
to "tune down" the page size when we list catalog entries. This pull
request introduces a `--catalog-entries-api-page-size` flag (and a
`CATALOG_ENTRIES_API_PAGE_SIZE` environment variable) that can be used
to customise the page size for this API call.