catalog-importer: 2.5.0 (#166)
We're releasing catalog-importer 2.5.0, which includes the ability to
have `schema_only` attributes in your importer config. The attributes on
catalog entries marked as 'schema_only' will not be updated when the
importer runs, and can be updated via the dashboard instead.

This is only additive functionality, and is not a breaking change -- all
current functionality will remain the same. You can read more about the
functionality in our example reference.jsonnet.