ONC-7399 Include generated enum types when pruning
The `--prune` logic was only looking at 'base' types (i.e., those
directly configured in an output), and not any enum types that are
inferred from this.

I've added a `.AllOutputTypes()` helper on `Config` to simplify a bunch
of loops over "every type in the config", which were correct but
duplicative in the create/update-schema loops.