Make --quiet a global flag
Like --debug, this should configure every command, and suppress
non-error output across all commands.

I've also introduced `ALWAYS_OUT` which prints output even when
`--quiet` is enabled - e.g. if the whole point of the command is to
produce that output.

Finally, `--quiet` is ignored if you're running `sync --dry-run`, since
the whole point of a dry run is to print things to the console!