Feature: Support use_name_as_identifier, icon, color for catalog type sync (#259)
Hey!

We noticed that we could not set `use_name_as_identifier` or `color` to
our catalog types, managed by catalog-importer.

Although the properties are generated in the client and available in the
API, they are not allowed fields in the JSON.

This PR adds `use_name_as_identifier`, `color` and `icon` fields.

We've tested the PR with our own sync process, and are happy with the
results: no issues when JSON is good, correct errors when icon/color is
invalid.

Note: we are mostly interested in `use_name_as_identifier`, I'll be
happy to split `icon` and `color` and the enum handling in another PR if
this can help getting the use_name_as_identifier part merged faster!

_PR code co-authored with AI_